+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## unreleased
 - `WebSockex.start_link` will no longer cause the calling process to exit on
   connection failure and will return a proper error tuple instead.
+- Change `WebSockex.Conn.RequestError` to `WebSockex.RequestError`.
+- Add `handle_connect_failure` to be invoked after initiating a connection
+  fails. Fixes #5
 
 ## 0.1.2
 - Rework how disconnects are handled which should improve the